requirements.txt
Flask
pytz
requests
python-dotenv
colorama

